requirements.txt
psycopg2-binary
PyYAML==3.13
requests
